# Fix: Performance tests require release mode for GitHub Actions