--- a/.claude/settings.local.json
+++ b/.claude/settings.local.json
@@ -76,7 +76,8 @@
       "Bash(CARGO_TERM_COLOR=always RUSTFLAGS=\"-D warnings\" CARGO_INCREMENTAL=0 cargo clippy --all-targets --all-features -- -D warnings)",
       "Bash(gh pr view:*)",
       "Bash(gh pr comment:*)",
-      "Bash(RUSTFLAGS=\"-D warnings\" cargo clippy --all-targets --all-features -- -D warnings)"
+      "Bash(RUSTFLAGS=\"-D warnings\" cargo clippy --all-targets --all-features -- -D warnings)",
+      "Bash(RUSTFLAGS=\"-D warnings\" cargo check --workspace --all-targets)"
     ],
     "deny": []
   }