--- a/crates/bingo-core/Cargo.toml
+++ b/crates/bingo-core/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 edition = "2024"
 
 [features]
-default = []
+default = ["arena-alloc"]
 arena-alloc = ["bumpalo"]
 parallel = ["rayon"]