--- a/fix_validation.md
+++ b/fix_validation.md
@@ -1,1 +1,0 @@
-# Performance Test Fix Validation