… .github/workflows/ci.yml
… .github/workflows/quality.yml
… ARCHITECTURE_EXTENSIBILITY.md
… CLAUDE.md
… IMPLEMENTATION_PLAN.md
… MIGRATION_STRATEGY.md
… OPTIMIZATION_FRAMEWORK.md
… PLUGIN_ARCHITECTURE.md
… PROJECT_COMPLETION_SUMMARY.md
… SPECS.md