## Summary

- Fix GitHub Actions workflows to run performance scaling tests in release mode
- Separate performance tests from regular unit tests in quality workflow
- Resolves CI/CD failures where debug mode caused 10x performance degradation

## Root Cause Analysis

Performance scaling tests were failing in GitHub Actions because they were running in debug mode instead of release mode. Debug builds have 10x performance degradation compared to release builds:

- **Release mode**: 100K: 0.67s ✅, 500K: 2.32s ✅, 1M: 6.85s ✅  
- **Debug mode**: 100K: 6.34s ❌, 500K: 23.41s ❌, 1M: 58.16s ❌

## Changes Made

### .github/workflows/quality.yml
- Separated unit tests from performance tests
- Added explicit release mode for scaling validation tests
- Updated test execution to exclude performance tests from debug runs

### Validation
- All local quality checks pass with ZERO tolerance
- Performance targets met in release mode
- Memory usage within enterprise production limits

## Test Results
```
✅ 100K facts: 0.67s (target: <3s)
✅ 500K facts: 2.32s (target: <10s) 
✅ 1M facts: 6.85s (target: <30s)
```

🚀 Generated with [Claude Code](https://claude.ai/code)